pyproject.toml
[tool.poetry]
name = "gordy's retrieval finetune harness"
version = "0.0.1"
description = "simple interface for finetuning document retrievers"
authors = ["Gordy"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.7,<4"
gradio = "^3.41.2"
pandas = "^2.1.0"
hnswlib = "^0.7.0"
torch = "^2.0.1"
datasets = "^2.14.4"
transformers = "^4.32.1"
matplotlib = "^3.7.2"
tqdm = "^4.66.1"
wandb = "^0.15.10"
faiss-cpu = "^1.7.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
